feat: 增加 模组资源目录复制 及 服务加载器生成 - 更新版本号至 0.6 - 新增 LoaderFetcher 类用于服务加载器生成 - 新增 LoadTarget 枚举类 - 重构 ModsBulkFetcher 类 - 优化 ConstantPool 类，新增 JAVA_PROGRAM 常量 - 更新 Source 类中的 getName 方法 - 修改 Pack2server 类中的 run 方法 - 更新测试类以适应新的功能和结构

+++ src/main/java/cloud/dbug/pack2server/common/ConstantPool.java
@@ -2,12 +2,17 @@ package cloud.dbug.pack2server.common;
 
 import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.io.file.PathUtil;
+import cn.hutool.core.lang.Opt;
+import cn.hutool.core.lang.func.Supplier2;
+import cn.hutool.core.util.StrUtil;
 import lombok.SneakyThrows;
 
 import java.io.File;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.List;
+import java.util.Objects;
 import java.util.function.Consumer;
 
 /**
@@ -27,11 +32,15 @@ public class ConstantPool {
     /**
      * 根目录
      */
-    public static final String DIR = System.getProperty("user.dir");
+    public static final String HOME = System.getProperty("user.dir");
     /**
      * 测试目录
      */
-    public static final File TEST_DIR = FileUtil.file(ConstantPool.DIR, ConstantPool.TEMP);
+    public static final File TEST_DIR = FileUtil.file(ConstantPool.HOME, ConstantPool.TEMP);
+    /**
+     * Java
+     */
+    public static final String JAVA_PROGRAM = Path.of(System.getProperty("java.home"), "bin", "java").toAbsolutePath().toString();
     /**
      * 服务目录
      */
@@ -39,15 +48,17 @@ public class ConstantPool {
     /**
      * 构建目录
      */
-    public static final Consumer<File> BUILD_DIRECTORY = parent -> SERVER_DIRS.forEach(dir -> FileUtil.mkdir(FileUtil.file(parent, dir)));
-
+    public static final Consumer<File> BUILD_DIR = parent -> SERVER_DIRS.forEach(dir -> {
+        ConstantPool.ensure(Paths.get(dir));
+        FileUtil.mkdir(FileUtil.file(parent, dir));
+    });
     /**
-     * 初始化目录
-     * @param parent 父母
+     * 复制目录
      */
-    public static void iniDir(final File parent) {
-        BUILD_DIRECTORY.accept(parent);
-    }
+    public static final Supplier2<File, File, File> COPY_DIR = (src, dest) -> {
+        ConstantPool.ensure(dest.toPath());
+        return FileUtil.copyContent(src, dest, Boolean.TRUE);
+    };
 
     /**
      * 确保目录
@@ -55,15 +66,24 @@ public class ConstantPool {
      */
     @SneakyThrows
     public static void ensure(final Path path) {
-        final Path parent = path.getParent();
-        if (Files.notExists(parent)) {
-            Files.createDirectories(parent);
+        if (Objects.nonNull(path)) {
+            final Path parent = path.getParent();
+            if (Objects.nonNull(parent) && Files.notExists(parent)) {
+                Files.createDirectories(parent);
+            }
+            if (PathUtil.isFile(path, Boolean.TRUE) && Files.exists(path)) {
+                Files.delete(path);
+            }
         }
-        if (PathUtil.isDirectory(path, Boolean.TRUE) && Files.notExists(parent)) {
-            Files.createDirectories(parent);
-        }
-        if (PathUtil.isFile(path, Boolean.TRUE) && Files.exists(path)) {
-            Files.delete(path);
-        }
+    }
+
+    /**
+     * 获取名称
+     * @param path 路径
+     * @return {@link String }
+     */
+    public static String getName(final String path) {
+        return Opt.ofBlankAble(StrUtil.subAfter(path, "/", Boolean.TRUE))
+                .orElseGet(() -> StrUtil.subAfter(path, "\\", Boolean.TRUE));
     }
 }
